+++ lib/api/objects/movable_object.cc
@@ -15,7 +15,7 @@ namespace objects {
 
 MovableObject::MovableObject(
     const Kind kind, const MovableObjectOpts options,
-    const std::vector<std::pair<int, int>>& hit_box_vertices)
+    const std::vector<std::pair<double, double>>& hit_box_vertices)
     : Object(kind,
              {.is_hit_box_active = options.is_hit_box_active,
               .should_draw_hit_box = options.should_draw_hit_box},
@@ -24,8 +24,8 @@ MovableObject::MovableObject(
       velocity_y_(options.velocity_y) {}
 
 MovableObject::MovableObject(const Kind kind, const MovableObjectOpts options,
-                             const std::pair<int, int> hit_box_center,
-                             const uint32_t hit_box_radius)
+                             const std::pair<double, double> hit_box_center,
+                             const double hit_box_radius)
     : Object(kind,
              {.is_hit_box_active = options.is_hit_box_active,
               .should_draw_hit_box = options.should_draw_hit_box},
